VerifyToken déplacer dans 'helper' pour require dans routes
