+++ routes/tickets.js
@@ -6,7 +6,7 @@ const router = express.Router()
 const cors = require('cors');
 const connection = require('../helpers/db.js');
 const bodyParser = require('body-parser');
-const jwt = require('jsonwebtoken')
+const verifyToken = require('../helpers/verifyToken')
 
 
 router.use(cors());
@@ -17,26 +17,7 @@ router.use(bodyParser.urlencoded({
 
 
 
-function verifyToken(req, res, next) {
-  const bearerHeader = req.headers.authorization
-  if (typeof bearerHeader !== 'undefined') {
-    const bearer = bearerHeader.split(' ')
-    const bearerToken = bearer[1]
-    req.token = bearerToken
-    jwt.verify(bearerToken, 'HPI_secretKey', (err, payload) => {
-      if (err) {
-        console.log('token invalid')
-        res.sendStatus(401)
-      }
-      else {
-        req.user = payload
-        next()
-      }
-    })
-  } else {
-    res.sendStatus(403)
-  }
-}
+
 
 
 // GET //
